requirements.txt
Flask
flask-cors
python-dotenv
pymongo
pandas
scikit-learn
